+++ CHANGELOG.md
@@ -27,6 +27,7 @@ See [keep a changelog] for information about writing changes to this log.
 - Change event images into own entity
 - Added image download and processing
 - Added geocoder services
+- Moved fixtures feeds into local filesystem
 
 [keep a changelog]: https://keepachangelog.com/en/1.1.0/
 [unreleased]: https://github.com/itk-dev/event-database-imports/compare/main...develop